fix: authentication added on routes

+++ server/graphql/resolvers/booking.js
@@ -1,35 +1,44 @@
-const Event = require('../../model/event');
-const Booking = require('../../model/booking');
-const {transformBooking, transformEvent } = require("./utils");
+const Event = require("../../model/event");
+const Booking = require("../../model/booking");
+const { transformBooking, transformEvent } = require("./utils");
 
 module.exports = {
-  bookings: async () => {
+  bookings: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error("Unauthenticated!");
+    }
     try {
       const bookings = await Booking.find();
-      return bookings.map(booking => {
+      return bookings.map((booking) => {
         return transformBooking(booking);
       });
     } catch (err) {
       throw err;
     }
   },
-  bookEvent: async args => {
+  bookEvent: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error("Unauthenticated!");
+    }
     const fetchedEvent = await Event.findOne({ _id: args.eventId });
     const booking = new Booking({
-      user: '5c0fbd06c816781c518e4f3e',
-      event: fetchedEvent
+      user: req.userId,
+      event: fetchedEvent,
     });
     const result = await booking.save();
     return transformBooking(result);
   },
-  cancelBooking: async args => {
+  cancelBooking: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error("Unauthenticated!");
+    }
     try {
-      const booking = await Booking.findById(args.bookingId).populate('event');
+      const booking = await Booking.findById(args.bookingId).populate("event");
       const event = transformEvent(booking.event);
       await Booking.deleteOne({ _id: args.bookingId });
       return event;
     } catch (err) {
       throw err;
     }
-  }
+  },
 };
